Create a shorter alias for returning view types
